# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Predic HSM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  uid: string;
  displayName: string;
  email: string;
  /** @nullable */
  photoURL?: string | null;
  walletBalance: number;
  /** @nullable */
  stream?: string | null;
  /** @nullable */
  username?: string | null;
  createdAt?: string;
  /** @nullable */
  lastDailyBonus?: string | null;
  totalBets?: number;
  totalWon?: number;
}

export interface ProfileInput {
  displayName: string;
  stream?: string;
  username?: string;
}

export interface ProfileUpdate {
  displayName?: string;
  stream?: string;
  username?: string;
}

export interface DailyBonusResult {
  claimed: boolean;
  /** @nullable */
  bonus?: number | null;
  /** @nullable */
  nextClaimAt?: string | null;
  message: string;
}

export type MarketStatus = typeof MarketStatus[keyof typeof MarketStatus];


export const MarketStatus = {
  open: 'open',
  locked: 'locked',
  resolved: 'resolved',
} as const;

export interface Market {
  id: string;
  title: string;
  description: string;
  category: string;
  status: MarketStatus;
  /** @nullable */
  outcome?: string | null;
  yesPool: number;
  noPool: number;
  /** @nullable */
  closesAt?: string | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  tags?: string[];
}

export interface MarketInput {
  title: string;
  description: string;
  category: string;
  closesAt?: string;
  imageUrl?: string;
  tags?: string[];
}

export type MarketResolveInputOutcome = typeof MarketResolveInputOutcome[keyof typeof MarketResolveInputOutcome];


export const MarketResolveInputOutcome = {
  YES: 'YES',
  NO: 'NO',
} as const;

export interface MarketResolveInput {
  outcome: MarketResolveInputOutcome;
}

export type BetSide = typeof BetSide[keyof typeof BetSide];


export const BetSide = {
  YES: 'YES',
  NO: 'NO',
} as const;

export type BetStatus = typeof BetStatus[keyof typeof BetStatus];


export const BetStatus = {
  active: 'active',
  locked: 'locked',
  won: 'won',
  lost: 'lost',
} as const;

export interface Bet {
  id: string;
  marketId: string;
  marketTitle?: string;
  userId: string;
  side: BetSide;
  amountPaid: number;
  sharesOwned: number;
  status: BetStatus;
  /** @nullable */
  payout?: number | null;
  placedAt: string;
}

export type BetInputSide = typeof BetInputSide[keyof typeof BetInputSide];


export const BetInputSide = {
  YES: 'YES',
  NO: 'NO',
} as const;

export interface BetInput {
  side: BetInputSide;
  amount: number;
}

export interface LeaderboardEntry {
  uid: string;
  displayName: string;
  /** @nullable */
  username?: string | null;
  walletBalance: number;
  totalWon?: number;
  totalBets?: number;
  rank: number;
}

export type GetLeaderboardPeriod = typeof GetLeaderboardPeriod[keyof typeof GetLeaderboardPeriod];


export const GetLeaderboardPeriod = {
  'all-time': 'all-time',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export type GetLeaderboardSort = typeof GetLeaderboardSort[keyof typeof GetLeaderboardSort];


export const GetLeaderboardSort = {
  richest: 'richest',
  'most-bets': 'most-bets',
  'most-won': 'most-won',
} as const;

export interface WalletAdjustInput {
  amount: number;
  reason?: string;
}

export type ListMarketsParams = {
category?: string;
status?: string;
};

export type GetLeaderboardParams = {
period?: GetLeaderboardPeriod;
sort?: GetLeaderboardSort;
};