# lib/api-zod/src/generated/types/bet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Predic HSM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BetSide } from './betSide';
import type { BetStatus } from './betStatus';

export interface Bet {
  id: string;
  marketId: string;
  marketTitle?: string;
  userId: string;
  side: BetSide;
  amountPaid: number;
  sharesOwned: number;
  status: BetStatus;
  /** @nullable */
  payout?: number | null;
  placedAt: string;
}