# lib/api-zod/src/generated/types/dailyBonusResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Predic HSM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyBonusResult {
  claimed: boolean;
  /** @nullable */
  bonus?: number | null;
  /** @nullable */
  nextClaimAt?: string | null;
  message: string;
}