# lib/api-zod/src/generated/types/getLeaderboardSort.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Predic HSM API specification
 * OpenAPI spec version: 0.1.0
 */

export type GetLeaderboardSort = typeof GetLeaderboardSort[keyof typeof GetLeaderboardSort];


export const GetLeaderboardSort = {
  richest: 'richest',
  'most-bets': 'most-bets',
  'most-won': 'most-won',
} as const;