# lib/api-zod/src/generated/types/marketInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Predic HSM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketInput {
  title: string;
  description: string;
  category: string;
  closesAt?: string;
  imageUrl?: string;
  tags?: string[];
}